chive_web/requirements.txt
Django==3.2.12
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-oauth-toolkit==1.7.0
django-redis==5.2.0
django-storages==1.12.3
djangorestframework==3.12.4
django-crispy-forms==1.14.0
drf-extensions==0.7.1
gevent==21.12.0
gunicorn==20.1.0
ipython==7.31.0
psycopg2-binary==2.9.3
pre-commit==2.17.0
channels==3.0.4
celery==5.2.6
channels-redis==3.4.0
beautifulsoup4==4.10.0
python-telegram-bot==13.11
sqlalchemy==1.4.35
lxml==4.8.0
django-crispy-forms==1.14.0

